Разбивает редьюсер на части wip
