+++ src/store/store.js
@@ -1,4 +1,4 @@
 import {createStore} from "redux";
-import {reducer} from "./reducer";
+import reducer from "./reducer/reducer";
 
 export default createStore(reducer, window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__());
